+++ packages/shared/src/types/index.ts
@@ -1,5 +1,4 @@
 export * from './api'
-export * from './boost'
 export * from './catalog'
 export * from './categories'
 export * from './categoryConfiguration'
@@ -31,6 +30,7 @@ export * from './sourceFieldOptionLabels'
 export * from './sourceFieldOptions'
 export * from './sourceFields'
 export * from './tabs'
+export * from './textFieldTags'
 export * from './tree'
 export * from './user'
 export * from './categorySortingOption'
