+++ packages/shared/src/types/textFieldTags.ts
@@ -1,0 +1,11 @@
+export interface IDataTags {
+  label: string
+  id: number
+}
+
+export interface ITextFieldTags {
+  label?: string
+  id: string
+  labelIsAll?: string
+  data: IDataTags[]
+}
